+++ composer.json
@@ -5,12 +5,14 @@
         "php": "^7.1.3",
         "ext-ctype": "*",
         "ext-iconv": "*",
+        "ausi/slug-generator": "^1.1",
         "symfony/console": "^4.1",
         "symfony/flex": "^1.0",
         "symfony/form": "^4.1",
         "symfony/framework-bundle": "^4.1",
         "symfony/lts": "^4@dev",
         "symfony/orm-pack": "^1.0",
+        "symfony/validator": "^4.1",
         "symfony/yaml": "^4.1"
     },
     "require-dev": {
