+++ config/orm/Article.orm.yml
@@ -1,6 +1,6 @@
-App\Entity\Article:
+App\Domain\Model\Article:
     type: entity
-    repositoryClass: App\Symfony\Repository\ArticleRepository
+    repositoryClass: App\Application\Repository\ArticleRepository
     table: article
     uniqueConstraints:
         unique_slug:
