+++ config/packages/doctrine.yaml
@@ -14,5 +14,5 @@ doctrine:
                 is_bundle: false
                 type: yml
                 dir: '%kernel.project_dir%/config/orm'
-                prefix: 'App\Entity'
+                prefix: 'App\Domain\Model'
                 alias: App
