+++ config/packages/framework.yaml
@@ -1,4 +1,5 @@
 framework:
+    csrf_protection: false
     secret: '%env(APP_SECRET)%'
     #default_locale: en
     #csrf_protection: true
