+++ config/packages/test/validator.yaml
@@ -1,0 +1,3 @@
+framework:
+    validation:
+        not_compromised_password: false
