+++ config/routes.yaml
@@ -1,16 +1,16 @@
 list_articles:
     path: /articles
     methods: [GET]
-    controller: App\Controller\ArticleController::list
+    controller: App\Application\Controller\ArticleController::list
 
 read_article:
     path: /articles/{slug}
     methods: [GET]
-    controller: App\Controller\ArticleController::read
+    controller: App\Application\Controller\ArticleController::read
     requirements:
-        slug: '[a-z\-]+'
+        slug: '[a-z0-9\-]+'
 
 create_article:
     path: /articles
     methods: [POST]
-    controller: App\Controller\ArticleController::create
+    controller: App\Application\Controller\ArticleController::create
