+++ config/services.yaml
@@ -1,33 +1,2 @@
-parameters:
-
-services:
-    _defaults:
-        autowire: true
-        autoconfigure: true
-        public: false
-
-    App\:
-        resource: '../src/*'
-        exclude: '../src/{Entity,Migrations,Tests,Kernel.php}'
-
-    App\Controller\:
-        resource: '../src/Controller'
-        tags: ['controller.service_arguments']
-
-    App\Command\Bus\SimpleBus:
-        arguments:
-            - '@service_container'
-
-    App\Command\Handler\ReadArticle:
-        public: true
-        arguments:
-            - '@App\Repository\ArticleRepository'
-
-    App\Command\Handler\ListArticles:
-        public: true
-        arguments:
-            - '@App\Repository\ArticleRepository'
-
-    App\EventListener\RequestListener:
-        tags:
-            - { name: kernel.event_listener, event: kernel.request }
+imports:
+    - { resource: services/ }
