+++ config/services/event_listeners.yaml
@@ -1,0 +1,4 @@
+services:
+    App\Application\EventListener\RequestListener:
+        tags:
+            - { name: kernel.event_listener, event: kernel.request }
