+++ config/services/vendors.yaml
@@ -1,0 +1,2 @@
+services:
+    Ausi\SlugGenerator\SlugGenerator:
