--- src/Command/Command.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace App\Command;
-
-interface Command
-{
-}
