--- src/Command/CommandResult.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App\Command;
-
-interface CommandResult
-{
-    /**
-     * @return mixed
-     */
-    public function getResult();
-}
