--- src/Command/CreateArticle.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace App\Command;
-
-class CreateArticle implements Command
-{
-    /**
-     * @var string
-     */
-    public $title;
-
-    /**
-     * @var string
-     */
-    public $content;
-}
