--- src/Command/ListArticles.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace App\Command;
-
-class ListArticles implements Command
-{
-}
