--- src/Command/ReadArticle.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace App\Command;
-
-class ReadArticle implements Command
-{
-    /**
-     * @var string
-     */
-    public $slug;
-
-    public function __construct(string $slug)
-    {
-        $this->slug = $slug;
-    }
-}
