+++ src/Domain/Command/Bus/CommandBus.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Domain\Command\Bus;
+
+interface CommandBus
+{
+    /**
+     * @param mixed $command
+     *
+     * @return mixed
+     */
+    public function executeCommand($command);
+}
