+++ src/Domain/Command/ListArticles.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace App\Domain\Command;
+
+class ListArticles
+{
+}
