+++ src/Domain/Command/ReadArticle.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace App\Domain\Command;
+
+class ReadArticle
+{
+    /**
+     * @var string
+     */
+    private $slug;
+
+    public function __construct(string $slug)
+    {
+        $this->slug = $slug;
+    }
+
+    public function getSlug(): string
+    {
+        return $this->slug;
+    }
+}
