+++ src/Domain/Repository/ArticleRepository.php
@@ -1,0 +1,8 @@
+<?
+
+namespace App\Domain\Repository;
+
+interface ArticleRepository
+{
+    public function findAllPublished(): iterable;
+}
