--- src/Entity/Article.php
@@ -1,75 +1,0 @@
-<?php
-
-namespace App\Entity;
-
-use Doctrine\ORM\Mapping as ORM;
-use DateTimeInterface;
-use DateTimeImmutable;
-
-class Article
-{
-    /**
-     * @var int
-     */
-    private $id;
-
-    /**
-     * @var string
-     */
-    private $title;
-
-    /**
-     * @var string
-     */
-    private $slug;
-
-    /**
-     * @var string
-     */
-    private $content;
-
-    /**
-     * @var DateTimeInterface
-     */
-    private $dateCreation;
-
-    /**
-     * @var bool
-     */
-    private $draft;
-
-    public function __construct()
-    {
-        $this->dateCreation = new DateTimeImmutable();
-    }
-
-    public function getId(): ?int
-    {
-        return $this->id;
-    }
-
-    public function getTitle(): ?string
-    {
-        return $this->title;
-    }
-
-    public function getSlug(): ?string
-    {
-        return $this->slug;
-    }
-
-    public function getContent(): ?string
-    {
-        return $this->content;
-    }
-
-    public function getDateCreation(): ?DateTimeInterface
-    {
-        return $this->dateCreation;
-    }
-
-    public function isDraft(): ?bool
-    {
-        return $this->draft;
-    }
-}
