+++ symfony.lock
@@ -1,4 +1,7 @@
 {
+    "ausi/slug-generator": {
+        "version": "v1.1.0"
+    },
     "doctrine/annotations": {
         "version": "1.0",
         "recipe": {
@@ -178,9 +181,6 @@
     },
     "symfony/mime": {
         "version": "v5.0.5"
-    },
-    "symfony/orm-pack": {
-        "version": "v1.0.8"
     },
     "symfony/options-resolver": {
         "version": "v4.4.5"
